Revert "Add haxe.hxmlFiles for HXML auto-discovery"
This has one major issue: instead of being limited to the top-level directoy, HXML files can be discovered in subdirectories with this. It's not clear how cwd should be handled for this hxml files: some might expect the subdirectory to be the cwd (example: hxml files generated by Kha), others expect the workspace root as the cwd (example: hxml files generated by Lime).

Handling this flexibly would add quite a bit of complexity, which is probably not worth it considering HXML discovery is mainly a convenience feature. Perhaps we should revisit this in the future, but I don't think this should be part of vshaxe 1.9.0.