DisplayArguments: fix no provider being selected if saved one doesn't exist
This can happen if the user explicitly selects Lime from the dropdown, after which it's saved, and then disables or uninstalls the Lime extension. In that case, no provider would be selected, which effectively leaves that workspace in a broken state until mementos are cleared or Lime is enabled again.

To avoid this, we always select a provider if none has been selected yet.